# api/openapi.yaml
openapi: 3.1.0
info:
  title: kbdb API
  description: >
    API for tracking a personal mechanical keyboard collection: keyboards,
    switches, keycap sets, and the builds that combine them. Routes are
    nested under /v1/users/{userId}/... to support both a caller's own
    collection and items other users have explicitly shared with them —
    see the userId path parameter's description for the exact rules.
  version: "1.0.0"
  contact:
    url: https://mykeebs.info
  license:
    name: MIT
    url: https://github.com/rogueserenity/kbdb/blob/main/LICENSE
servers:
  - url: https://api.mykeebs.info
security:
  - BearerAuth: []
...
paths:
  /v1/users/{userId}/keyboards:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [keyboards]
      summary: List keyboards
      operationId: listKeyboards
      security:
        - {}
        - BearerAuth: []
      description: >
        If userId is the caller's own subject, lists all of the caller's keyboards. Otherwise, lists only that user's keyboards explicitly shared with the caller (possibly empty).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of keyboards.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyboardListPage"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [keyboards]
      summary: Create a keyboard
      operationId: createKeyboard
      description: >
        userId must be the caller's own subject; creating in another user's collection always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyboardInput"
      responses:
        "201":
          description: The created keyboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyboard"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /v1/users/{userId}/keyboards/{keyboardId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/KeyboardId"
    get:
      tags: [keyboards]
      summary: Get a keyboard
      operationId: getKeyboard
      security:
        - {}
        - BearerAuth: []
      description: >
        Returns the keyboard if it belongs to the caller, or if its owner has shared it with the caller. Returns 404 (not 403) if the keyboard does not exist, is not owned by the caller, and is not shared with the caller — this avoids revealing that an unshared item exists.
      responses:
        "200":
          description: The requested keyboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyboard"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    put:
      tags: [keyboards]
      summary: Update a keyboard
      operationId: updateKeyboard
      description: >
        userId must be the caller's own subject; updating another user's keyboard always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyboardInput"
      responses:
        "200":
          description: The updated keyboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyboard"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    delete:
      tags: [keyboards]
      summary: Delete a keyboard
      operationId: deleteKeyboard
      description: >
        userId must be the caller's own subject; deleting another user's keyboard always returns 404.
        on_delete controls what happens if the keyboard is still referenced by a build: "block" (the
        default) fails with 409 and lists the blocking build ids; "cascade" deletes the keyboard and
        every build that references it, returning their ids; "detach" deletes the keyboard regardless,
        leaving referencing builds with a dangling keyboard field.
      parameters:
        - name: on_delete
          in: query
          schema:
            type: string
            enum: [block, cascade, detach]
            default: block
          description: How to handle a keyboard still referenced by one or more builds.
      responses:
        "200":
          description: Deleted via cascade; lists the builds that were also deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CascadeDeleteResult"
        "204":
          description: Deleted (block with no references, or detach).
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          description: Blocked by one or more referencing builds (on_delete=block, the default).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
  /v1/users/{userId}/keyboards/{keyboardId}/images/{imageId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/KeyboardId"
      - name: imageId
        in: path
        required: true
        schema:
          type: string
        description: The image_id returned when the image was added.
    delete:
      tags: [keyboards]
      summary: Remove an image from a keyboard
      operationId: deleteKeyboardImage
      description: >
        userId must be the caller's own subject; removing an image from
        another user's keyboard always returns 404.
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /v1/users/{userId}/switches:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [switches]
      summary: List switches
      operationId: listSwitches
      security:
        - {}
        - BearerAuth: []
      description: >
        If userId is the caller's own subject, lists all of the caller's switches. Otherwise, lists only that user's switches explicitly shared with the caller (possibly empty).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of switches.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwitchListPage"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [switches]
      summary: Create a switch
      operationId: createSwitch
      description: >
        userId must be the caller's own subject; creating in another user's collection always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchInput"
      responses:
        "201":
          description: The created switch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Switch"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An IdP-issued OIDC JWT (Authorization Bearer <token>).
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
      description: >
        The IdP subject of the item's owner — the `sub` claim of an access
        token, and the `user_id` field of a profile. If this matches the
        authenticated caller's own subject, the request is against their own
        collection. If it names a different user, the request is against
        that user's collection: reads return only items that user has
        explicitly shared with the caller (or, for a specific item, only if
        that item is shared with the caller — otherwise a 404, not a 403,
        to avoid revealing that an item exists but isn't shared); writes
        (create/update/delete) always fail with a 404 in this case, since a
        caller can never modify another user's items. Not a PII/security
        concern to expose in a URL — an IdP subject is an opaque identifier,
        not derived from email or any other identifying attribute.
    KeyboardId:
      name: keyboardId
      in: path
      required: true
      schema:
        type: string
      description: Server-generated identifier for the keyboard.
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response's next_cursor.
  responses:
    ProblemResponse:
      description: An error, formatted per RFC 9457 (Problem Details for HTTP APIs).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFoundResponse:
      description: The requested resource does not exist, or does not belong to the authenticated user.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details.
      properties:
        type:
          type: string
          format: uri
          example: https://mykeebs.info/errors/validation-error
        title:
          type: string
          example: Validation Error
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        invalid_params:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              reason:
                type: string
              allowed_values:
                type: array
                items:
                  type: string
        blocking_build_ids:
          type: array
          description: >
            Present on a 409 response caused by the item still being
            referenced by one or more builds. Lists the ids of the blocking
            builds.
          items:
            type: string
      required: [type, title, status]
    CascadeDeleteResult:
      type: object
      description: >
        Returned by a delete call made with on_delete=cascade, listing the
        ids of every build that was deleted alongside the requested item.
      properties:
        deleted_build_ids:
          type: array
          items:
            type: string
      required: [deleted_build_ids]
    Money:
      type: number
      format: double
      description: A price in the collection's tracking currency.
      minimum: 0
      maximum: 100000000000000
      example: 123.45
...
    Visibility:
      type: string
      enum: [public, authenticated, private]
      description: >
        Who can read this item via its GET routes (create/update/delete
        always require the caller to be the item's owner, regardless of
        visibility). "public": readable with no authentication at all.
        "authenticated": readable by any signed-in user, but not
        anonymous callers. "private": readable only by the owner. There
        is no per-user allowlist/sharing mechanism beyond these three
        tiers — visibility is not scoped to specific other users.
    PurchaseInput:
      type: object
      description: >
        Full purchase-lifecycle tracking, used by keyboards and keycap set
        kits (which get separate order/delivery dates and status).
      additionalProperties: false
      properties:
        vendor:
          type: string
          description: Open vocabulary — validated against the "vendor" lookup at request time.
        price:
          $ref: "#/components/schemas/Money"
        order_date:
          type: string
          format: date
        delivery_date:
          type: string
          format: date
        order_status:
          type: string
          description: Open vocabulary — validated against the "order_status" lookup at request time.
...
    KeyboardDesign:
      type: object
      additionalProperties: false
      properties:
        top_case:
          $ref: "#/components/schemas/MaterialColor"
        bottom_case:
          $ref: "#/components/schemas/MaterialColor"
        weight:
          $ref: "#/components/schemas/MaterialColor"
        plates:
          type: array
          items:
            type: string
            description: Open vocabulary — validated against the "keyboard_plate_material" lookup at request time.
    KeyboardPCB:
      type: object
      additionalProperties: false
      properties:
        thickness:
          type: number
          format: double
          example: 1.6
        firmware:
          type: string
          description: Open vocabulary — validated against the "keyboard_pcb_firmware" lookup at request time.
        assembly:
          type: string
          description: Open vocabulary — validated against the "keyboard_pcb_assembly_type" lookup at request time.
        connectivity:
          type: string
          description: Open vocabulary — validated against the "keyboard_pcb_connectivity_type" lookup at request time.
    KeyboardInput:
      type: object
      required: [brand, name, visibility]
      additionalProperties: false
      properties:
        brand:
          type: string
          minLength: 1
          pattern: "\\S"
        name:
          type: string
          minLength: 1
          pattern: "\\S"
        size:
          type: string
          description: Open vocabulary — validated against the "keyboard_size" lookup at request time.
        layout:
          type: string
          description: >
            Open vocabulary — validated against the "keyboard_layout" lookup at
            request time, including that the chosen layout is valid for
            the keyboard's size.
        design:
          $ref: "#/components/schemas/KeyboardDesign"
        pcb:
          $ref: "#/components/schemas/KeyboardPCB"
        purchase:
          $ref: "#/components/schemas/PurchaseInput"
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
...
    Keyboard:
      type: object
      required: [id, brand, name]
      properties:
        id:
          type: string
        brand:
          type: string
        name:
          type: string
        size:
          type: string
        layout:
          type: string
        design:
          $ref: "#/components/schemas/KeyboardDesign"
        pcb:
          $ref: "#/components/schemas/KeyboardPCB"
        purchase:
          $ref: "#/components/schemas/Purchase"
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
          description: >
            Who can read this keyboard. Only ever present for the
            keyboard's owner - no preference can expose it to another
            caller.
        images:
          type: array
          items:
            $ref: "#/components/schemas/KeyboardImage"
          description: >
            Managed via the dedicated image endpoints below, not via
            PUT on the keyboard itself — see POST/DELETE
            .../keyboards/{keyboardId}/images[/{imageId}].
    KeyboardSummary:
      type: object
      description: Reduced shape returned in list responses.
      properties:
        id:
          type: string
        brand:
          type: string
        name:
          type: string
        size:
          type: string
        layout:
          type: string
        order_status:
          type: string
        image:
          $ref: "#/components/schemas/KeyboardImage"
          description: First entry of images, if any (a freshly-minted presigned URL, same as on the full Keyboard resource).
        price:
          $ref: "#/components/schemas/Money"
          description: >
            What the keyboard cost. Present for the owner only if their
            show_price_to_me preference is set, and for any other caller
            only if the owner's show_price_to_others is.
        currency:
          $ref: "#/components/schemas/Currency"
          description: >
            The owner's display currency, for labelling price.
            Present exactly when price is.
        visibility:
          $ref: "#/components/schemas/Visibility"
          description: >
            Who can read this keyboard. Only ever present for the
            keyboard's owner - no preference can expose it to another
            caller.
    KeyboardListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/KeyboardSummary"
        next_cursor:
          type: [string, "null"]
...
    SwitchInput:
      type: object
      required: [brand, name, type, visibility]
      additionalProperties: false
      properties:
        brand:
          type: string
          minLength: 1
          pattern: "\\S"
        manufacturer:
          type: string
        name:
          type: string
          minLength: 1
          pattern: "\\S"
        type:
          type: string
          description: Open vocabulary — validated against the "switch_type" lookup at request time.
        pins:
          type: integer
          example: 5
          description: PCB mount pin count.
        factory_lubed:
          type: boolean
        material:
          $ref: "#/components/schemas/SwitchMaterial"
        force:
          $ref: "#/components/schemas/SwitchForce"
        spring:
          $ref: "#/components/schemas/SwitchSpring"
        purchase:
          $ref: "#/components/schemas/SwitchPurchaseInput"
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
...
    Switch:
      type: object
      required: [id, brand, name, type]
      properties:
        id:
          type: string
        brand:
          type: string
        manufacturer:
          type: string
        name:
          type: string
        type:
          type: string
        pins:
          type: integer
          example: 5
          description: PCB mount pin count.
        factory_lubed:
          type: boolean
        material:
          $ref: "#/components/schemas/SwitchMaterial"
        force:
          $ref: "#/components/schemas/SwitchForce"
        spring:
          $ref: "#/components/schemas/SwitchSpring"
        purchase:
          $ref: "#/components/schemas/SwitchPurchase"
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
          description: >
            Who can read this switch. Only ever present for the
            switch's owner - no preference can expose it to another
            caller.
        image:
          $ref: "#/components/schemas/SwitchImage"
          description: >
            Managed via the dedicated image endpoint, not via PUT on
            the switch itself — see POST/DELETE
            .../switches/{switchId}/image.
...
    SwitchListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SwitchSummary"
        next_cursor:
          type: [string, "null"]
...